# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital SaaS API",
    "description": "Public Swagger documentation for Hospital SaaS Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://YOUR_PRODUCTION_DOMAIN",
      "description": "Production (Replace manually)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "security": [{ "BearerAuth": [] }],
        "summary": "List Users",
        "responses": { "200": { "description": "User list" } }
      }
    },
    "/api/users/create-doctor": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Create Doctor",
        "responses": { "200": { "description": "Doctor created" } }
      }
    },
    "/api/users/create-nurse": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Create Nurse",
        "responses": { "200": { "description": "Nurse created" } }
      }
    },
...
    "/api/prescriptions": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Add Prescription",
        "responses": { "200": { "description": "Prescription added" } }
      }
    },
    "/api/prescriptions/{id}": {
      "put": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Update Prescription",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Prescription updated" } }
      },
      "delete": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Delete Prescription",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Prescription deleted" } }
      }
    },
...
  }
}